feat: add authentication verify in account controller

+++ src/main/java/br/com/finance/controllers/AccountController.java
@@ -31,14 +31,15 @@ public class AccountController {
   private AccountService accountService;
   
   @GetMapping
-  public ResponseEntity<List<AccountResponseDto>> getAccountsByUserAuthenticated(){
+  public ResponseEntity<List<AccountResponseDto>> getAccountsByUser(){
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
     if(authentication != null && authentication.getPrincipal() instanceof User){
       User user = (User) authentication.getPrincipal();
-      Long id = user.getId();
-      List<Account> accounts = this.accountService.getAccountsByUser(id);
+      Long userId = user.getId();
+      List<Account> accounts = this.accountService.getAccountsByUser(userId);
       List<AccountResponseDto> accountsDto = accounts.stream().map(AccountResponseDto::new).toList();
+
       return ResponseEntity.ok().body(accountsDto);
     }
     
@@ -47,10 +48,21 @@ public class AccountController {
 
   @GetMapping("/{id}")
   public ResponseEntity<AccountResponseDto> getAccountById(@PathVariable(value = "id") Long id){
-    var account = this.accountService.findAccountById(id);
-    var accountResponseDto = new AccountResponseDto(account);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-    return ResponseEntity.ok().body(accountResponseDto);
+    if(authentication != null && authentication.getPrincipal() instanceof User){
+      User user = (User) authentication.getPrincipal();
+      Long userId = user.getId();
+      Account account = this.accountService.findAccountById(id);
+
+      if(account != null && account.getUserId().equals(userId)){
+        return ResponseEntity.ok().body(new AccountResponseDto(account));
+      }
+
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
+    
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }
  
   @PostMapping
@@ -60,7 +72,7 @@ public class AccountController {
     if(authentication != null && authentication.getPrincipal() instanceof User){
       User user = (User) authentication.getPrincipal();
       Long id = user.getId();
-      var account = this.accountService.createAccount(request.toAccount(), id);
+      Account account = this.accountService.createAccount(request.toAccount(), id);
 
       return ResponseEntity.status(HttpStatus.CREATED).body(new AccountResponseDto(account));
     }
@@ -69,22 +81,39 @@ public class AccountController {
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<AccountResponseDto> updateAccount(@PathVariable Long id, @RequestBody AccountRequestDto accountRequestDto){
+  public ResponseEntity<AccountResponseDto> updateAccount(@PathVariable(value = "id") Long idAccount, @RequestBody AccountRequestDto accountRequestDto){
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    User user = (User) authentication.getPrincipal();
-    Long userId = user.getId();
     
-    try {
-        Account accountUpdate = this.accountService.updateAccount(userId, accountRequestDto.toAccount()); 
+    if(authentication != null && authentication.getPrincipal() instanceof User){
+      User user = (User) authentication.getPrincipal();
+      Long userId = user.getId();
+      Account accountUpdate = this.accountService.updateAccount(userId, idAccount, accountRequestDto.toAccount());
+
+      if(accountUpdate != null && accountUpdate.getUserId().equals(userId)){
         return ResponseEntity.ok().body(new AccountResponseDto(accountUpdate));
-    } catch (IllegalArgumentException e) {
-        return ResponseEntity.notFound().build();
+      }
+
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
+
+    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
   }  
 
   @DeleteMapping("/{id}")
   public ResponseEntity<Object> deleteAccount(@PathVariable(value = "id") Long id){
-    this.accountService.deleteAccount(id);
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    
+    if(authentication != null && authentication.getPrincipal() instanceof User){
+      User user = (User) authentication.getPrincipal();
+      Long userId = user.getId();
+      Account account = this.accountService.findAccountById(id);
+      
+      if(account != null && account.getUserId().equals(userId)){
+        this.accountService.deleteAccount(id);
+      }
+
+      return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+    }
 
     return ResponseEntity.noContent().build();
   }
